+++ tool/util.h
@@ -48,7 +48,7 @@ size_t read_data(unsigned char*, size_t, FILE*, enum enum_format);
 void dump_data(unsigned const char*, unsigned int, FILE*, bool, enum enum_format);
 unsigned long set_length(unsigned char*, unsigned long);
 unsigned long get_length(const unsigned char*, unsigned long*);
-bool has_valid_length(const unsigned char*, unsigned long);
+bool has_valid_length(const unsigned char*, ptrdiff_t);
 int get_curve_name(int);
 X509_NAME *parse_name(const char*);
 unsigned char get_algorithm(EVP_PKEY*);
